added order search function in backend. added order.js in api route for customers to search for past orders

+++ migrations/20220815162625-customers.js
@@ -23,7 +23,7 @@ exports.up = function(db) {
     email:{type:'string',length:50,notNull:true},
     password:{type:'string',length:100,notNull:true},
     shipping_address:{type:'string',length:100,notNull:true},
-    postal_code:{type:'smallint',unsigned:true,notNull:true},
+    postal_code:{type:'int',unsigned:true,notNull:true},
     mobile_number:{type:'int',unsigned:true,notNull:true},
     datetime_created:{type:'datetime', notNull:true},
     datetime_last_modified:{type:'datetime', notNull:true}
